Refactor resolve_handshake_pattern() and name() functions
resolve_handshake_pattern is removed in favor of methods on the
HandshakePattern enum. resolve_handshake_pattern used a set of output
parameters, which are typically a code smell in Rust. HandshakePattern
now has separate methods which provide the handshake name, pre-message
patterns and message patterns as static values, without copying to an
out param.

The name() method on the CipherStateType, DhType, CipherType, and
HashType have been changed to return a &'static str instead of taking an
&mut [u8] out-param. Same rational for the resolve_handshake_pattern
change.

It may be possible to split this commit and do the two changes
separately, but I found they fit together pretty naturally.

The end result is reduced complexity in HandshakeState. HandshakeState
has been reduced in size by roughly 80 bytes.