Refactor DhType trait
This commit adds associated types to the DhType to represent the public
and private key types. The intent is that the types will always be
static-length arrays. The DhType trait is now immutable (the keys can't
be mutated after creating the concrete DhType).

Additionally, the HandshakeState struct now owns its DH keys instead of
holding references to them. The DH keys are relatively cheap to copy (64
bytes) and removing the references means the lifetime param is one step
closer to being unneeded. If copying the 64 bytes of keys becomes a
performance issue, the DhType can always be implemented for Rc<T> and
Arc<T> where T: DhType.

Finally, the HandshakeState struct holds the Dh keys in Options where
appropriate, instead of having separate bool flags.